Move SupportedProtocolVersions to sui-types, rename AuthorityCapabilities -> AuthorityCapabilitiesV1
