+++ crates/sui-types/src/lib.rs
@@ -79,6 +79,7 @@ pub mod storage;
 pub mod sui_sdk2_conversions;
 pub mod sui_serde;
 pub mod sui_system_state;
+pub mod supported_protocol_versions;
 pub mod traffic_control;
 pub mod transaction;
 pub mod transfer;
