+++ crates/sui-types/src/supported_protocol_versions.rs
@@ -1,0 +1,39 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+pub use sui_protocol_config::{Chain, ProtocolConfig, ProtocolVersion};
+
+/// Models the set of protocol versions supported by a validator.
+/// The `sui-node` binary will always use the SYSTEM_DEFAULT constant, but for testing we need
+/// to be able to inject arbitrary versions into SuiNode.
+#[derive(Serialize, Deserialize, Debug, Clone, Copy, Hash, PartialEq, Eq)]
+pub struct SupportedProtocolVersions {
+    pub min: ProtocolVersion,
+    pub max: ProtocolVersion,
+}
+
+impl SupportedProtocolVersions {
+    pub const SYSTEM_DEFAULT: Self = Self {
+        min: ProtocolVersion::MIN,
+        max: ProtocolVersion::MAX,
+    };
+
+    /// Use by VersionedProtocolMessage implementors to describe in which range of versions a
+    /// message variant is supported.
+    pub fn new_for_message(min: u64, max: u64) -> Self {
+        let min = ProtocolVersion::new(min);
+        let max = ProtocolVersion::new(max);
+        Self { min, max }
+    }
+
+    pub fn new_for_testing(min: u64, max: u64) -> Self {
+        let min = min.into();
+        let max = max.into();
+        Self { min, max }
+    }
+
+    pub fn is_version_supported(&self, v: ProtocolVersion) -> bool {
+        v.as_u64() >= self.min.as_u64() && v.as_u64() <= self.max.as_u64()
+    }
+}
